pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "dspy>=3.0.2",
    "fastapi>=0.116.1",
    "google-cloud-secret-manager>=2.24.0",
    "marimo>=0.15.2",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.1.1",
    "python-json-logger>=3.3.0",
    "requests>=2.32.5",
    "ruff>=0.12.11",
    "uvicorn>=0.35.0",
    "yoyo-migrations>=9.0.0",
]
